+++ core-ng-search-test/src/test/java/core/framework/search/module/TestSearchConfigTest.java
@@ -1,8 +1,5 @@
 package core.framework.search.module;
 
-import core.framework.search.impl.ElasticSearchImpl;
-import core.framework.search.impl.MockElasticSearch;
-import core.framework.test.module.TestModuleContext;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -20,9 +17,8 @@ class TestSearchConfigTest {
     }
 
     @Test
-    void createElasticSearch() {
-        ElasticSearchImpl search = config.createElasticSearch(new TestModuleContext());
-        assertThat(search).isInstanceOf(MockElasticSearch.class);
+    void configureLogger() {
+        config.configureLogger();
 
         assertThat(System.getProperty("log4j.configurationFactory")).isNotNull();
     }