+++ core-ng-search/src/main/java/core/framework/search/SearchException.java
@@ -9,8 +9,4 @@ public class SearchException extends RuntimeException {
     public SearchException(String message) {
         super(message);
     }
-
-    public SearchException(Throwable cause) {
-        super(cause);
-    }
 }