+++ core-ng-search/src/main/java/core/framework/search/module/SearchConfig.java
@@ -18,50 +18,41 @@ import java.time.Duration;
 public class SearchConfig extends Config {
     ElasticSearchImpl search;
     private ModuleContext context;
-    private String host;
     private boolean typeAdded;
 
     @Override
     protected void initialize(ModuleContext context, String name) {
+        if (name != null) throw Exceptions.error("search does not support multiple instances, name={}", name);
         this.context = context;
+        configureLogger();
         search = createElasticSearch(context);
         context.beanFactory.bind(ElasticSearch.class, null, search);
     }
 
     @Override
     protected void validate() {
-        if (host == null) throw new Error("search host must be configured");
+        if (search.host == null) throw new Error("search host must be configured");
         if (!typeAdded)
-            throw Exceptions.error("elasticsearch is configured but no type added, please remove unnecessary config");
+            throw new Error("elasticsearch is configured but no type added, please remove unnecessary config");
     }
 
-    ElasticSearchImpl createElasticSearch(ModuleContext context) {
-        System.setProperty("log4j2.loggerContextFactory", ESLoggerContextFactory.class.getName());
-        ElasticSearchImpl search = new ElasticSearchImpl();
+    public void host(String host) {
+        search.host = host;
+    }
+
+    private ElasticSearchImpl createElasticSearch(ModuleContext context) {
+        var search = new ElasticSearchImpl();
         context.startupHook.add(search::initialize);
         context.shutdownHook.add(ShutdownHook.STAGE_10, timeout -> search.close());
         return search;
     }
 
-    public void host(String host) {
-        setHost(host);
-        this.host = host;
-    }
-
-    void setHost(String host) {
-        search.host(host);      // es requires host must be resolved, skip for unit test
-    }
-
-    public void sniff(boolean sniff) {
-        search.sniff = sniff;
+    void configureLogger() {
+        System.setProperty("log4j2.loggerContextFactory", ESLoggerContextFactory.class.getName());
     }
 
     public void slowOperationThreshold(Duration threshold) {
         search.slowOperationThreshold = threshold;
-    }
-
-    public void timeout(Duration timeout) {
-        search.timeout = timeout;
     }
 
     public <T> void type(Class<T> documentClass) {