+++ log-processor/src/test/java/core/log/job/CleanupOldIndexJobTest.java
@@ -1,14 +1,13 @@
 package core.log.job;
 
+import core.framework.search.ClusterStateResponse;
 import core.framework.search.ElasticSearch;
-import core.framework.search.ElasticSearchIndex;
 import core.log.service.IndexService;
-import org.elasticsearch.cluster.metadata.IndexMetaData;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
-import java.util.List;
+import java.util.Map;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -32,10 +31,12 @@ class CleanupOldIndexJobTest {
 
     @Test
     void cleanup() {
-        var indices = List.of(index("action-2017.10.01", IndexMetaData.State.OPEN),
-                index("action-2017.10.30", IndexMetaData.State.CLOSE),
-                index("action-2017.11.01", IndexMetaData.State.OPEN));
-        when(elasticSearch.indices()).thenReturn(indices);
+        var state = new ClusterStateResponse();
+        state.metadata = new ClusterStateResponse.Metadata();
+        state.metadata.indices = Map.of("action-2017.10.01", index(ClusterStateResponse.IndexState.OPEN),
+                "action-2017.10.30", index(ClusterStateResponse.IndexState.CLOSE),
+                "action-2017.11.01", index(ClusterStateResponse.IndexState.OPEN));
+        when(elasticSearch.state()).thenReturn(state);
 
         job.cleanup(LocalDate.of(2017, 11, 8));
 
@@ -43,10 +44,9 @@ class CleanupOldIndexJobTest {
         verify(elasticSearch).closeIndex("action-2017.11.01");
     }
 
-    private ElasticSearchIndex index(String index, IndexMetaData.State state) {
-        ElasticSearchIndex elasticSearchIndex = new ElasticSearchIndex();
-        elasticSearchIndex.index = index;
-        elasticSearchIndex.state = state;
-        return elasticSearchIndex;
+    private ClusterStateResponse.Index index(ClusterStateResponse.IndexState state) {
+        ClusterStateResponse.Index index = new ClusterStateResponse.Index();
+        index.state = state;
+        return index;
     }
 }